Rename browse controller to features/queries controller
