+++ config/locales/en.yml
@@ -411,11 +411,13 @@ en:
       telephone_link: "Call %{phone_number}"
       colour_preview: "Colour %{colour_value} preview"
       email_link: "Email %{email}"
-    query:
-      title: "Query Features"
-      introduction: "Click on the map to find nearby features."
-      nearby: "Nearby features"
-      enclosing: "Enclosing features"
+  features:
+    queries:
+      show:
+        title: "Query Features"
+        introduction: "Click on the map to find nearby features."
+        nearby: "Nearby features"
+        enclosing: "Enclosing features"
   old_elements:
     index:
       node:
